+++ src/main/java/net/equj65/jedisbench/runnner/OperationRunner.java
@@ -1,0 +1,9 @@
+package net.equj65.jedisbench.runnner;
+
+import java.util.concurrent.Callable;
+
+/**
+ * Created by ryozo on 2016/02/08.
+ */
+public abstract class OperationRunner implements Callable<Void> {
+}
